+++ include/path_tracer/ray.h
@@ -1,7 +1,7 @@
 #ifndef path_tracer_ray_h
 #define path_tracer_ray_h
 
-#include "vec3.h"
+#include "path_tracer/vec3.h"
 
 class Ray {
 public:
@@ -11,7 +11,7 @@ public:
     Ray() {}
     Ray(const Vec3& o, const Vec3& d) : origin(o), direction(d) {}
 
-    Vec3 point_at_parameter(float t) const {
+    Vec3 point_at_parameter(double t) const {
         return origin + t * direction;
     }
 };
