+++ include/path_tracer/sphere.h
@@ -1,31 +1,31 @@
 #ifndef path_tracer_sphere_h
 #define path_tracer_sphere_h
 
-#include "hitable.h"
-#include "material.h"
+#include "path_tracer/hitable.h"
+#include "path_tracer/material.h"
 
 class Sphere: public Hitable {
 public:
     Vec3 center;
-    float radius;
+    double radius;
     Material* material;
     
     Sphere() {}
-    Sphere(Vec3 c, float r, Material* m) : center(c), radius(r), material(m) {}
+    Sphere(Vec3 c, double r, Material* m) : center(c), radius(r), material(m) {}
     virtual bool hit(
-        const Ray& r, float t_min, float t_max, HitRecord& rec) const;
+        const Ray& r, double t_min, double t_max, HitRecord& rec) const;
 };
 
 bool Sphere::hit(
-    const Ray& ray, float t_min, float t_max, HitRecord& record) const
+    const Ray& ray, double t_min, double t_max, HitRecord& record) const
 {
     Vec3 oc = ray.origin - center;
-    float a = dot(ray.direction, ray.direction);
-    float b = dot(oc, ray.direction);
-    float c = dot(oc, oc) - radius * radius;
-    float discriminant = b * b - a * c;
+    double a = dot(ray.direction, ray.direction);
+    double b = dot(oc, ray.direction);
+    double c = dot(oc, oc) - radius * radius;
+    double discriminant = b * b - a * c;
     if (discriminant > 0.0f) {
-        float temp = (-b - sqrt(discriminant)) / a;
+        double temp = (-b - sqrt(discriminant)) / a;
         if (temp < t_max && temp > t_min) {
             record.t = temp;
             record.p = ray.point_at_parameter(record.t);
